backend/requirements.txt
dnspython==1.16.0
flask==1.1.1
flask-mongoengine==0.9.5
flask-restx==0.1.1
flask-script==2.0.6
pymongo==3.7.2
python-dotenv==0.10.3
werkzeug==0.16.1 # flask-restx 0.1.1 has requirement werkzeug<=0.16.1
pymongo[srv]
python-dateutil==2.6.1